build_predict_1/requirements.txt
Flask
tensorflow
Pillow
cassandra-driver